Fix missing dependencies (as seen from the api)
Issue #17 describes a problem with gem 
dependencies missing when viewed through the api.
These same dependencies are visible from the gem
and from the rubygems.org dependency list.

The fix is actually quite simple: the version_id
was not being captured in the case where new
versions are inserted for gems. As dependencies
are associated with both the gem and the version
(associations set up in the Rails app) the 
API returns no dependencies when the version is
missing.

Also added the built_at attribute to be the same
as that used in rubygems.org - it was causing
postgres constraint violation in development.