Add default ttl, 15 min to gem cache
Not sure how, but we are getting some gem cache out-of-sync.
Probably some race condition, or a something else.
If we add a ttl, it will fix the issue, even if that doesnt make me happy.

We can play if the TTL, after we deploy this, we can watch the number of cache miss,
and lower or increase this number