versions.list file never changes a bit
With this change, we group the gems by created_at before name, so we can
have different entries for gems names, but never for created_at.

This causes the versions.list file to never change a bit, no matter if
recreated or when adding the gems dynamically at the bottom.