Move response bodies to compact_index gem
The response body is also shared by rubygems.org and bundler-api. So it
is natural to move this code to compact-index gem.

This patch does not remove the spec format response. I think it is
healty to keep some basic test on the response format as well, so we can
do more complete tests on the compact-index itself.