quote ETags per rfc2616, parse If-None-Match too
Turns out that If-None-Match can be a comma separated list of ETags, and
if the ETags aren’t quoted, they’re expected to be interpreted as a
timestamp rather than an opaque identifier. This change quotes ETags on
responses, and requires quoted ETags in the If-None-Match header, as
well as allowing a list of quoted ETags to be supplied for matching.