Add a caching strategy and a storage
For now this is quite simple, given a folder, we can create a storage
which is capable of getting a CachedGem instance.

This can either exist or not, and can be used to save both headers and
the content of the file.

The storage is tested, the caching strategy has only 1 test that checks
that given a gem name, it will fetch the gem, store it, and return this
cached data.

Next steps will be to test further and to actually call rubygems to
fetch the gem and the headers from there.